Use stale data on failure
Summary:
The native autoloader returns `true` if we successfully update the map and `false` otherwise. The question is... what should we do if we return `false` and haven't updated our map successfully?

Usually we still have a perfectly good map which will probably suffice to run the request. When we use stale data, we'll only crash if we encounter a class that was added or moved since the last time we updated the map.

I could (should?) leave this to a config option, but is anyone going to say they want their requests to crash more often? The only reason I think people would want this is if they want a higher reliability guarantee in the middle of some code (because they're communicating with multiple endpoints and don't have some transactional guarantee to make the communication all-or-nothing). But even then, native autoloader code doesn't run in production. And if you work in WWW, you should already be aware that your code can fail at any time...

Reviewed By: dneiter

Differential Revision: D17347137

fbshipit-source-id: f4517ad94483faff6ab2eaa1ac773abadc3bd27b