Make IDE file sync work with saved files
Summary:
This diffstack syncs up the behavior of LocalIndex with TrieIndex, and moves all our integration tests / unit tests to run on LocalIndex mode.

This diff cleans up "test_ide_sync" slightly so that it reuses behavior from the integration_test_base and provides support for file saving.

A future diff on this stack will switch autocomplete to run on the LocalIndex mode, which will help ensure that the new autocomplete modes are exercised by our test suites.

Reviewed By: 2BitSalute

Differential Revision: D17352786

fbshipit-source-id: cf29dad46bc2b67f2dc1b025bd7214243ef0678d