Delete Typing_print.suggest
Summary:
Typing_print.suggest was added in the early, early days (2014 or earlier) days of Hack. It's intention was to suggest a type if one was missing in a partial file. This code has bit rotten over the years and I couldn't come up with a case where it actually gets triggered in code.

There are a few other instances where it is used, which I can assume because it is the only function in Typing_print that takes only the type as an argument. I replaced those calls with other print functions to the best of my knowledge. Suggestions on testing those would be appreciated, but given how broken Typing_print.suggest is I doubt it will be a worst experience.

Reviewed By: andrewjkennedy

Differential Revision: D17323099

fbshipit-source-id: 0440784550a22cced1ffc601a6ca2725fbaf34f6