Create type aliases locl_ty/decl_ty
Summary:
In preparation for removal of `decl ty` in favor of `Aast.hint` I am replacing direct references of `decl ty` and `locl ty` with alias `decl_ty` and `locl_ty`. This way I can change what those aliases are backed by more easily.

At the same time I renamed the phases `decl` and `locl` to `decl_phase` and `locl_phase` respectively. This is so when rebasing on this diff you will not refer to the phases directly.

Another part of this was changes to `fun_type` to change it's type parameter from taking a phase to taking a ty. This is also needed to help with the eventual change. Because of this I've made alias for `decl fun_type` so it is referred to as `decl_fun_type` instead.

Reviewed By: kmeht

Differential Revision: D17344295

fbshipit-source-id: fa14584bb4d84ce298eaae9c74740253c0e541fe