Eliminate simple usage of ty as a GADT
Summary:
In order to get rid of `decl ty` we need to eliminate cases where we are using it as GADT, i.e. the function operates on both `decl ty` and `locl ty`.

There are not that many functions that exist which do this. In this diff I take on a few simple cases.

In typing_generic we only ever use the code on `decl_ty`.

In type_serialization_identity_check it is only used on a `locl_ty`

For `Typing_make_type.nullable` I split it into two separate functions `nullable_decl` and `nullable_locl`.

Reviewed By: losvald

Differential Revision: D17346866

fbshipit-source-id: 000990fa44313bc4f2d7cbf449ddc05c6e0df448