Split type_visitor into decl_type_visitor and locl_type_visitor
Summary:
The Type_visitor operates on types on either the decl or locl phase. For the majority of visitors they only ever operated on a single phase. I made decl_type_visitor and locl_type_visitor and switched users of the visitor to use either one or the other. All but two visitors were trivially migrated. I will clean those up in a follow up diff.

As a result of this there are some instances where code was written to handle cases that will never occur. I've removed those functions instead of trying to make them work preserve existing behavior.

Reviewed By: vassilmladenov

Differential Revision: D17348551

fbshipit-source-id: 77c9fe10f7e4b328e0eecf43a80dcb9ab7114c73