Create a fresh variable for missing type parameters
Summary:
This is an experiment to create new fresh variables for missing type parameters when we localize a type hint.

This diffs reuses the logic that can be found when creating a new object in order to create new fresh variables and add the corresponding type constraints when some type parameters are missing from a type hint.

Reviewed By: andrewjkennedy

Differential Revision: D16782995

fbshipit-source-id: 8e6637f185a640f35f8b247b8a9525c2882d005c