proper control flow typing for goto
Summary:
This uses a map of continuations per label. A `goto l` stores the context in a continuation indexed by `l`, and a label restores this context.

We don't do fix point computation in case the label is before the goto, but I don't think PPL has such usages. So still unsound, but less.

Reviewed By: andrewjkennedy

Differential Revision: D14672663

fbshipit-source-id: 2f2b34c34404c0688c61a0cf3b21f10c37d96886