Any generic constraint holds in an inconsistent environment
Summary:
Consider the following example:

```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

class A {
  public function f(): void where int as string {}
}

class B extends A {}
```

The definitions of class `A` and its method `f` are accepted by the type checker as well-typed.  Trying to call `f` on an instance of `A` causes a type error as expected.  However, the definition of `B` is rejected with an error that class `B` does not correctly implement all required members, more specifically that `f` has the wrong type because `int` is not a subtype of `string`.  This is odd: I would `B` to be accepted.

I think the issue here boils down to the distinction between *asserting* and *decomposing* subtyping constraints.  Type checking of `f` involves checking that the function type `(function(): void where int as string)` is a valid override for itself.  This, in turn, involves checking that the where constraint `int as string` holds for the subtype *assuming* the constraints that are imposed by the supertype hold.  It seems that `int as string` must imply `int as string`.  However, the former (the constraint belonging to the supertype) is added to `tpenv` after decomposition, and since it simplifies to an empty disjunction (because this constraint is unsatisfiable), nothing is added to `tpenv`.  It feels like in such cases, when decomposing an unsatisfiable constraint, we should not just drop it but somehow record that our context now contains a false proposition and hence *ex falso quodlibet*.

Here I've tried doing that in a naive way: by augmenting the type parameter environment with a boolean field `consistent` that is set to `false` when an unsatisfiable constraint (a false assumption) is added to the environment.  If the current `tpenv` is inconsistent, then any generic type parameter constraint is considered to just hold.

I'm not sure the above reasoning is correct and welcome suggestions.  On the one hand, it feels that inclusion of an unsatisfiable constraint into the set of assumptions should make any code type check under this set of assumptions (because this code must be unreachable -- there should be no instantiation of generic type parameters with concrete types at run time that would satisfy all the constraints).  On the other hand, if I add say `int as string` to the set of assumptions, this doesn't immediately allow me to prove an arbitrary subtyping proposition `t1 <: t2`.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D17093966

fbshipit-source-id: 3076487693cbebfa4108fe15b14f4f34aff8a72b