Fix reified generics double-free
Summary:
When we pass at least two varargs arguments on the stack, the second varargs argument will be in the reified generics local slot. If we use tvSet to set the reified generics, it'll dec-ref the varargs argument, causing a double-free.

There may be a related bug that causes a leak. In non-repo mode, we create and pass reified generics whenever we call a generic function with an explicit type parameter list. However, if the function doesn't actually use reified generics, it'll overwrite this list with a stack element. I don't see it getting freed.

Since leaking memory won't cause heap corruption, we can land this fix now and take some more time to investigate it.

Reviewed By: ricklavoie

Differential Revision: D17352274

fbshipit-source-id: c879255226e7667f1af241b9a177602f729078eb