Switch LSP test framework from `lib2to3` to `libcst`
Summary:
It turned out that the way we were doing things in `lib2to3` was flagging the wrong code in our tests, since it was hard to determine the true range of a function call. This was super-confusing during debugging since it could point to unrelated LSP requests having issues.

Additionally, the `lib2to3` API is kind of bad, and not all of it is recognized by Pyre. Switch to Instagram's `libcst` since the API is better and it supports Pyre pretty well.

I also filed https://github.com/Instagram/LibCST/issues/42 along the way.

Differential Revision: D16969801

fbshipit-source-id: 8a5c05cdfd66cf7928310325565eb5c37fbd91b3