Stop dropping certain messages in LSP test cases
Summary:
Before this commit, there was some obscure bug where we dropped certain messages. It's unclear exactly when it reproduces, but it caused a signatureHelp test case to erroneously pass while I was examining it.

We have a special test function to wait for `hh_server` to be initialized. This is accomplished by sending a dummy request and waiting for a non-error response. We previously accomplished this by sending the request over and over again, and waiting for the response to be not-an-error.

Suppose that we send the dummy request 5 times and get 3 failures followed by 2 successes. We'll actually exit the initialization function after the first success, regardless of the fact that there was another one queued up. Other parts of the code have subtle dependencies on how many responses they read in, so this was causing those parts of the code to read in a useless response, which just silently overwrites the one we have.

We fix this by using new primitives to ensure that these initialization requests are paired one-for-one with their responses. This exposes an issue in the existing signature help test case. We also take a more declarative approach where we wait for *some* message matching the criteria to appear, rather than mandate that it be the most recent message (since they could be interleaved due to circumstances out of our control).

Differential Revision: D16969803

fbshipit-source-id: bbba7bb78b94ac5dece02cde55090a8f930cb85c