Have serverRewriteTypeParamsType use locl_type_visitor
Summary: Think this was reverted with a bad merge. I should note that the filtering of Tnothing this is supposed to do was dead code before since this code is only ever called with `locl_ty`. I deleted the code and I'll leave it for poechsel to reimplement with proper testing.

Differential Revision: D17372524

fbshipit-source-id: df4707662d605bf84bad1e3ffad11a2fd6e2f65b