Refactor Type_validator to use a decl_type_visitor
Summary:
The type_validator uses type_visitor to do validation of enforceable types and to check if a type can be reified. This was done by doing some validation on a decl_ty, running this validation when localization was done, and then doing additional validation on the locl_ty.

For the current usage of type_validator, we should really only need to validate against a decl_ty. The only reason localization is necessary is to handle expansion of aliases. So that is what I do. I change type_validator to use decl_type_visitor and implement `on_tapply` to handle expansion of aliases, as well as dispatch out to other visitors `on_class`, `on_enum`, `on_newtype`, and `on_alias`. I also centralized the logic of handling type constants since much of it was duplicated between validators.

As a result I could delete the dty validator code from Typing_phase and also delete type_visitor since this was the last place using it.

Reviewed By: vassilmladenov

Differential Revision: D17379638

fbshipit-source-id: 611b21628d7b627c3bbad4b636a30a243c459463