Derive Ocamlvalue for all oxidized types
Summary: When the Ocamlvalue procedural macro was added, only types relevant to AST-lowering were whitelisted for deriving this trait. This may have been because the procedural macro didn't cover unit structs, so it was more expedient to derive it only for types which don't have any. Since adding support for unit structs is pretty straightforward, I suggest that we derive Ocamlvalue for all oxidized types instead. This allows us to allocate decls on the OCaml heap from Rust.

Reviewed By: 2BitSalute

Differential Revision: D17407603

fbshipit-source-id: b39b2e6784eeb6bae2f80d716d16a3794118c96f