Save global_tvenv when done typechecking each file
Summary:
This diff adds the ability to save the global tvenv generated by the typechecker to the disk. The implementation is in three steps:
- Modification of `Typing.fun_def` and `Typing.class_def` so that they can return the list of `global_tvenv` generated while typing methods and functions
- Concatenation of these lists to obtain a list of all generated `global_tvenv` for each file
- Marshalling of these lists when global inference is activated

We are currently marshalling the lists of `global_tvenv` to the disk as it is easier than adding a new thread-safe data structure. In the future the saved global_tvenvs will be read by `hh_server` in order to build a global constraint graph.

Reviewed By: andrewjkennedy

Differential Revision: D17162193

fbshipit-source-id: 8162e4af30baa26458ce903d43c24303be783827