Fix hhbc exception with 0 length property names
Summary:
A deep, dark corner case causes exceptions in hhbc (and hhvm). It requires a property with 0 length (see attached test cases to see how to get that). That yields an obscure error saying index out of bounds at line -1.

This fixes the error and dumps the stack trace in `hh_single_compile` cli mode so the next person spends less time that I did hunting this down.

Reviewed By: alexeyt

Differential Revision: D17400946

fbshipit-source-id: 687e35ecac483084b070b21865b02abcbb66402c