Add new iter alias classes
Summary:
Specialized iterator code looks a lot better if we clearly describe the memory effects of storing to each iterator field and let load- and store-elim do their thing.

We now have four iterator fields: the base, type, pos, and end. The end is useful because we can precompute it as a pointer for tombstone-free mixed arrays; for other arrays, it can still let us avoid a load of the array's header. In this diff, we extend AliasClass to understand all four iterator fields. The diff is very straightforward; the two only somewhat interesting things about it are:
1. We no longer do precise unions of two different iterator fields. Instead, we make IterBase { $id } | IterPos { $id } union to IterAll { $id }.
2. This diff also includes a small bugfix in store-elim to prevents us from sinking stores to iterator fields into inline frames. This bugfix has no effect now but is important in the specialized iterator stack.

Reviewed By: ricklavoie

Differential Revision: D17000147

fbshipit-source-id: a717650ad0fb51849c175eea4de773cf7e32c2eb