Correctly compute the expected key and value types if the expected collection type is mixed or nonnull
Summary:
Consider the following example:
```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

function expect_mixed(mixed $_): void {}

function test(int $a, string $b): void {
  expect_mixed(
    dict[
      'a' => $a,
      'b' => $b,
    ],
  );
}
```
Hack infers the type `dict<mixed, mixed>` for the argument of `expect_mixed`.  That's because when we do bidirectional type checking we incorrectly compute the expected types of keys and values from the expected type `mixed` of the dict: we return `mixed` for both even though the top type for keys is `arraykey`.

You might be surprised that we don't infer the more precise type `dict<string, (int | string)>` for the argument of `expect_mixed`.  However, Hack is free to pick any type for the expression as long as it satisfies the constraints imposed by the typing rules.  When the dict is constructed and immediately passed to a function, as in the example above, computing the precise type of the dict is wasteful and can hurt performance (we may need to compute and simplify a very large union if there are many, say hundreds of entries).

Reviewed By: andrewjkennedy

Differential Revision: D17419003

fbshipit-source-id: 81981a0293137fc21abcda2dffdd4d2220c6b775