Catch exceptions in saved_state_loader
Summary:
In the current implementation of saved_state_loader and futures, it is considered acceptable to throw an exception if the saved state cannot be found.

For the Sqlite search provider, which we intend to use for versions of hh_server outside of facebook, we do not want saved_states to crash the process if they cannot be found.

Let's put a catch around the attempt to use the saved state system.  This will ensure that any future developers who add exceptions to saved states do not break sqlite search provider.

Reviewed By: 2BitSalute

Differential Revision: D17404397

fbshipit-source-id: 1b180aa2719ec3bcd4a2b2a89ed836425b52171a