Enable coercion to const types, pessimize property types and class consts
Summary:
This uses the localization function from {D17353712} to carry the enforced local type for coercion in a global constant definition.
```
const int X = "4" as ~int;
```
This is useful for fixing opaque enums. It also lifts out the method-level pessimization setting to class-level, so now initializers to unenforceable types work properly as like types.

Reviewed By: Matt-Schellhas

Differential Revision: D17355045

fbshipit-source-id: 84aee4af4f89235f36d8aec801d557d9ededfb1e