Convert ocamlrep values to native Rust values
Summary:
While ocamlrep allows easy conversion of Rust values to OCaml values, there is currently no way of converting in the other direction. This might be an okay situation in the long term, but for now, we have OCaml modules whose output we would like to use in Rust, and it would be nice to have a means of using those outputs before a Rust replacement for those modules is ready. Additionally, we have existing logic for sending messages from an OCaml process to a Rust process, and a derived means of conversion would allow us to do away with the handwritten boilerplate for building those messages.

This change adds the `from_ocamlrep` method to the `OcamlRep` trait. It takes an `ocamlrep::Value` (which may have been allocated by ocamlrep, or may be a value on the OCaml heap--since their layout is the same, it does not matter which it is) and parses it as a value of type `Self` (the implementor of the trait). I've attempted to provide a descriptive error type for cases where that parse fails, in hopes that it will help us investigate any conversion failures that we might encounter in the future.

Instead of adding `from_ocamlrep` as a method on the OcamlRep trait, we could add a new trait instead. This would allow us to implement Rust-to-OCaml conversion without requiring us to implement OCaml-to-Rust conversion for some types (or vice-versa). I'm not sure that this is necessary, though (it isn't at this time, at least). It isn't possible to implement OCaml-to-Rust conversion for `&str` with the definition of `from_ocamlrep` as it is, but I don't think this is a major issue. If we found that string-copying was a bottleneck we wanted to avoid, we could take the route serde took--separate the Serialize and Deserialize traits, and parameterize Deserialize by the lifetime of the output. This seems like more effort than it's worth, though.

Reviewed By: 2BitSalute

Differential Revision: D17349385

fbshipit-source-id: 16c90fbb143304e95a826cf97e6a11c345f8200b