Check for errors in files containing __PPL attribute
Summary:
If a file contains `__PPL` attribute, we don't check for parser errors in it in HHVM. Some archeology reveals how we got here:

-  it was added in D7041109 (Feb 2018) as an emergency stopgap for some breakage (cc Arun, Andrew), with an ask for "coroutine experts to look this over post-land"
- in D8292584 (Jun 5, 2018), coroutine experts (Michael, Thomas) incorrectly blame it on jamesjwu who only refactored that code. There are some doubts about if it's needed / correct, but "for parity reasons, this should be left in in the meantime."
- in D8553911 (Jun 20, 2018), the same experts have the same conversation, and point out to the original diff as "something would break without it"

IMHO whatever reasons we had for that, they are probably long gone (let's wait for sandcastle test run to confirm).
 ---
BUT, while I have all you gathered here, I would like to address one more issues around coroutines:

We check (in hh_server, and after this diff, also in HHVM) the file before rewriting, but we never check the file AFTER it is processed. It is (at least theoretically) possible that there exists an input that causes the ppl rewriter / coroutine lowerer to generate garbage code that now goes straight into emitter without any checks, and could possibly generate garbage bytecode and crash everything in production. How worried are we about it?

The reason I ask is that we are very close to finishing porting the error checking behavior to Rust ({D16901289}), and I always assumed that rustifying coroutine code is prerequisite to shipping it (since I thought that we check the post-transformation code). But it turns out that it's not, and we could do the switch much sooner!

The only risk is that if we suddenly decide that we DO want to start checking the post-transformation code too, there will be no easy and quick way to do this (because the checking logic will operate only on Rust structs, without a way to go back from OCaml to Rust). It will require rewriting the coroutine code to Rust first, which we would like to punt on for now (though still get to it sometime in the next few weeks).

The problem scenarios seem very unlikely (less than 100 files, with no active development, without anything breaking despite problem existing for more than a year) but I'll ask anyway - is anyone strongly against this course of action?

For Michael/Thomas/Arun - who is actually used the PPL code in production? Can you point me to some people /diffs  who care about this code and would be upset if it breaks, so we can ask them what level of risk are they willing to tolerate?

Reviewed By: arun11111

Differential Revision: D17359661

fbshipit-source-id: f4f8ff5218539ea8912f6a8c4dc1fc7db9c74b68