Env support backtrack
Summary:
At several places lower env needs to create an env "scope" for children and throw it away later. This diff uses keeps "global" data in `Rc<RefCell<State>>`, so a new temporary env can be created by cloning.

A different approach is to implement lowerer as a `struct` and keep all "global" data in struct. I quickly tried it, it is not very friendly with high order functions, which is used in entire lowerer.

- every function needs to take two parameters, `this` and `env`.
- `this.foo` can't be passed to another function directly, a closure wrapper is required.

Reviewed By: losvald

Differential Revision: D17202367

fbshipit-source-id: 94cd3f126142edad9c812d171817fc59ff28c3b4