Allow projecting type constants off of generic type parameters
Summary:
At the moment, we don't allow projecting type constants off of generic type parameters as in
```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

interface HasFoo {
  abstract const type TFoo;
  public function getFoo(): this::TFoo;
}

function test<T as HasFoo>(T $x): T::TFoo {
  return $x->getFoo();
}
```
This produces an error:
```
T must be an identifier for a class, "self", or "this" (Naming[2061])
```
However, it *is* possible to use projections off of generics in where constraints, so the above can be written (a bit awkwardly) as follows:
```
<?hh
// Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.

interface HasFoo {
  abstract const type TFoo;
  public function getFoo(): this::TFoo;
}

function test<T as HasFoo, TFoo>(T $x): TFoo where TFoo = T::TFoo {
  return $x->getFoo();
}
```
The transformation of the former example into the latter looks pretty mechanical, which suggests that we might be able to lift the limitation that type constants can't be projected off of generics?

In fact, I simply tried to remove the check that a type constant access with a generic type parameter at the root occurs in a `where` constraint, and that seems to be enough to make the above example work.  I'd be curious to know if there are any soundness implications of this change.

Reviewed By: dlreeves

Differential Revision: D17344674

fbshipit-source-id: 0c0c4eb358b85b95ead3eb4897bb00c2cb4bd1f2