Avoid scanning nonexistent repositories
Summary:
This diff stack is intended to switch the default autocomplete provider to Sqlite.

Note that some tests simulate a fake repository with the root folder of `/`.  This means that the sqlite index builder might attempt to scan the root folder for a list of PHP files.  This diff also updates the behavior of the index builder to avoid scanning folders that do not have `.hhconfig` files.  As per Fred Emmott, all open source users have `.hhconfig` files at the root of their repository.

Reviewed By: 2BitSalute

Differential Revision: D17375982

fbshipit-source-id: 688dd2880ea3b0dca9ad2b08404dddae7fc85e1c