Delete --typed-full-fidelity-json command
Summary: The typed-full-fidelity-json client command was added more than a year ago, but from what I can tell has been broken for a majority of types. I am deleting this command because of this. This allows me to change Typing_print.to_json to only accept locl_ty. This code is actually tested since we check if it run trips in Type_serilaization_identity_check

Reviewed By: dabek

Differential Revision: D17441983

fbshipit-source-id: dbb15214655a646902146f952082f78f5463b6ca