Make Hhhbc_options' names consistent with HHVM Runtime
Summary:
Spotted a few inconsistencies while doing regular expressions for
porting Hhbc_options module to Rust. The options are stored in config
under names that are snake versions of HHVM runtime options,
(source of thruth), which is a lossless conversion.  However,
these outliers are both confusing and complicate Rust-OCaml interop:

- create_**inout**_wrapper_functions ("in" and "out" should be separate words)
- notice_on_**byref**_argument_typehint_violation (similar to the above)
- **enable_perf_logging** (should be: log_extern_compiler_perf)
- optimize_null_check (should be: optimize_null_check**s**)

Reviewed By: shiqicao

Differential Revision: D17407880

fbshipit-source-id: 431172a795d2dd48863e1c9a39d010e863432b65