Add support for 6th bytecode immediate
Summary:
This diff is pure boilerplate that allows us to have six immediates for an HHBC op. The motivation is to add an additional bit to LIter ops to track whether the base is ever changed; if we know the base is unchanged during iteration it is much easier to do a couple optimizations.

This diff is messy, because these numerical macros are used in a lot of different places. I have two alternatives to this diff that are partially complete, but I think this way is better than both:
1. Create a new "IterArgs" struct to store the immediates for the eight IterInit and IterNext ops. IterArgs would include an iterator ID, an offset, a value local ID, and an optional key local ID. Doing so would cut down the number of immediates for these ops from 3-5 to 1-3.
2. Replace these ONE, TWO, THREE macros with a var-args X macro. Use a functional FOR_EACH_VA helper in each usage of the X macro that evaluates a macro on each element of the var-args list, optionally with an index.

To see why 1 would make the code unnecessarily complicated, take a look at the three (!) structs encoding FCallArgs: FCallArgsBase (all fields except the offset), jit::FCallArgs (which uses Offset for the offset type), and op::FCallArgs (which uses BlockId for the offset type). I don't want three different IterArgs structs...

One alternative - call it 1' - is to include the iter ID, value local ID, and key local ID in IterArgs but to leave the offset out. This approach would be much simpler than putting the offset in.

I think 2 is also a good alternative. My main concerns with it are that a) the implementation of FOR_EACH_VA would be kind of complicated internally (making use of deferring recursive macro expansion between multiple preprocessor passes) and b) that macro would have to stay defined everywhere in our codebase instead of being isolated. But it will produce the cleanest usage code and will lift the number of immediates restriction completely.

Reviewed By: mofarrell

Differential Revision: D17399019

fbshipit-source-id: 8ff6f9941c52d756cab2d4f6e23d7676882c396b