Move Cargo.toml files to /cargo/ directory
Summary:
We inherit "multiple targets in single directory" from `buck`, which is incompatible with `Cargo`, which requires single crate per directory. To mitigate it, we created a bunch of directories (quiet a few of them, after recent splitting of the parser) that contain nothing but `Cargo.toml` file. I think they make it harder to browse codebase, since you see a folder with interesting name, think that there will be some code inside, while there isn't. I would prefer not to think and see them at all. In this diff I move them to a single directory.

I'm open to suggestions if someone has different feelings about it.

Reviewed By: shiqicao

Differential Revision: D17470419

fbshipit-source-id: f4fba59a8d5d98f6f1b00479bc448d62b07e4900