Multiply entities beyond necessity even more (force better build parallelism)
Summary:
Bad news everyone! My brilliant plan from {D17404542} actually doesn't work - it helps `dev` mode, but regresses `opt`. Looks like in `opt` mode it ends up forcing the compiler to do additional work that can't be re-used later on, increasing the general time it takes to compile.

This is a bit worrying, because it suggests that the compilation cost of generic can be O(uses) instead of O(unique uses). We'll see going forward how much it hurts.

Instead, in this diff I implement what I mention there as an alternative - splitting end crate into multiple targets too and ensuring that the actual functions that we are interested in will get compiled in parallel, and not the dummy stubs.

Reviewed By: shiqicao

Differential Revision: D17467988

fbshipit-source-id: 29eba7cabb0f88660066b99eeec7f1b35e1f107a