Add slab allocator
Summary:
The ocamlrep::Arena type takes after other Rust arenas (particularly, [arena::Arena](https://doc.rust-lang.org/1.1.0/arena/struct.Arena.html)) in that it is designed to be:

- a fast bump-allocator,
- for many values of the same lifetime,
- which is unbounded in size.

This is useful for implementing a method like `into_ocamlrep`, where it is convenient to treat allocation as infallible (made possible by the unbounded size).

However, to store ocamlrep values in cachelib, we need the storage for these values to have some different properties:

- The memory region must be contiguous.
- The memory region is not allocated or owned by us (it's allocated and owned by cachelib).
- We need the ability to fix up pointers when that memory region is moved to a different location by cachelib.

The unbounded size of Arenas requires the ability to allocate (also, the current implementation of growing results in Arenas being backed by non-contiguous memory chunks, but this could be fixed). This makes Arenas not suitable for the purpose of cachelib storage.

To solve this problem, this change introduces a new memory management structure named `Slab`. A `Slab` is a contiguous array of ocamlrep values prefixed with a few words of metadata. While many ocamlrep values are stored in a Slab, Slabs are designed to represent a tree of values with a single root (e.g., a single decl).

In order to fix up pointers when a Slab is moved, the first word of metadata in a Slab is the memory location where the Slab resided when all its pointers were last updated, known as the "base pointer". If the Slab's pointers are up-to-date, the base pointer will point to its own memory address. If the Slab has been moved, the base pointer and the Slab's current location in memory will differ, and the values in the Slab cannot be used without fixing up their pointers first (by adding the difference between the old base pointer and new base pointer to each). This fix-up operation is called "rebasing".

The second word of metadata in a slab is the offset (in words) of the Slab's root value from the beginning of the slab. In this diff, we only build slabs by recursively copying values, so we are able to easily guarantee that the root value is the first one allocated (and thus appears at the beginning of the data region). I've included this metadata in anticipation of the ability to build slabs from Arenas. We'd like to simply memcpy Arena chunks into a new slab rather than recursively traversing a value, but since Arena exposes a raw block-allocation API, we don't have a guarantee that an implementor of `into_ocamlrep` will allocate the root value first. As a result, we need to be able to handle the case where the root value occurs at some other offset within the slab.

The third word of metadata is a magic number, written only after the slab has been fully initialized (by writing a value tree into it). This is intended to be a sanity check--we will be reading slabs from cachelib as byte strings, so we can have no guarantee from the type system that those bytes are a valid slab.

Here's an example of what we store in a Slab, using the value `Some "a"`, with the slab rebased to address `0x00`:

| address | value                         | comment                      |
| ------- | ----------------------------- | ---------------------------- |
| 0x00    | 0x00                          | base pointer                 |
| 0x08    | 4                             | root value offset            |
| 0x10    | 0x51A851A8                    | magic number                 |
| 0x18    | Header { size: 1, tag: 0 }    | Some's tag is 0              |
| 0x20    | 0x30                          | pointer (base ptr + 6 words) |
| 0x28    | Header { size: 1, tag: 252 }, | STRING_TAG = 252             |
| 0x30    | 0x0600000000000061,           | "a"                          |

Reviewed By: jewelpit

Differential Revision: D17431857

fbshipit-source-id: 7241afb96749ca8553d90cf4f73fe9efbe5504e8