support building with Boost 1.65.1
Summary:
This unblocks targetting Ubuntu 18.04, which unblocks bumping the minimum GCC
for FB's open source C++ code.

We didn't find this out before as we explicitly install `libboost-foo-dev`, and if any of the ones
we depend on aren't present when we run cmake, we use the pinned version from hhvm-third-party. The
build environment list grew out of sync with cmake's list, so we were missing a required part of boost,
so we built a newer version.

On FB test infra, the ubuntu 18.04 image ended up having `libboost-all-dev` installed, so we got everything,
making us find this bug.

This should also make builds faster - both FB internal OSS tests, and real external builds.

This change is required because we pass it into `boost::apply_visitor()`, which in older versions, requires a const visitor

Reviewed By: yfeldblum

Differential Revision: D17489058

fbshipit-source-id: 16184f1200b0fb9faa2b2ada75affe0ecf4ee0d8